v0.4

+++ app/main/routes.py
@@ -6,11 +6,11 @@ from flask import render_template, flash, redirect, url_for, request,current_app
 from flask_login import current_user, login_required
 from flask_ckeditor import upload_success,upload_fail
 
-from app import db
-from app.models import User,Microblog,Blog,Category
-from app.utils import redirect_back
+from app.extensions import db
+from app.models import Microblog,Blog,Category
+from app.utils import redirect_back,allowed_file
 from app.main import bp
-from app.main.forms import EditProfileForm,MicroblogForm,BlogForm,CategoryForm
+from app.main.forms import MicroblogForm,BlogForm,CategoryForm
 
 
 @bp.before_request
@@ -25,13 +25,12 @@ def before_request():
 @login_required
 def index():
     page = request.args.get('page', 1, type=int)
-    categories = Category.query.filter_by(author=current_user).order_by(Category.id)
-    blogs = Blog.query.filter_by(author=current_user,is_released=True).order_by(Blog.timestamp.desc()).paginate(page, current_app.config['POSTS_PER_PAGE'],False)
-    next_url = url_for('main.index',page=blogs.next_num) \
-        if blogs.has_next else None
-    prev_url = url_for('main.index',page=blogs.prev_num) \
-        if blogs.has_prev else None
-    return render_template('index.html',categories=categories, blogs=blogs.items,next_url=next_url,prev_url=prev_url)
+    #pagination = Blog.query.filter_by(author=current_user,is_released=True).order_by(Blog.timestamp.desc()).paginate(page, current_app.config['POSTS_PER_PAGE'])
+    pagination = current_user.get_releasedblogs().paginate(page, current_app.config['POSTS_PER_PAGE'])
+    blogs = pagination.items
+    categories = current_user.get_categories()
+    links = current_user.get_links()
+    return render_template('index.html', blogs=blogs,categories=categories, links=links, pagination=pagination)
     
 
 @bp.route('/microblog/new',methods=['GET','POST'])
@@ -47,26 +46,20 @@ def new_microblog():
         return redirect(url_for('main.new_microblog'))
 
     page = request.args.get('page',1,type=int)
-    microblogs = current_user.followed_microblogs().paginate(
-        page,current_app.config['POSTS_PER_PAGE'],False)
-    next_url = url_for('main.new_microblog',page=microblogs.next_num) \
-        if microblogs.has_next else None
-    prev_url = url_for('main.new_microblog',page=microblogs.prev_num) \
-        if microblogs.has_prev else None
-    return render_template('microblog/new_microblog.html',title='Microblog',form=form,microblogs=microblogs.items,next_url=next_url,prev_url=prev_url)
+    pagination = current_user.followed_microblogs().paginate(
+        page,current_app.config['POSTS_PER_PAGE'])
+    microblogs = pagination.items
+    return render_template('microblog/new_microblog.html',form=form, microblogs=microblogs, pagination=pagination)
 
 
 @bp.route('/explore')
 @login_required
 def explore():
     page = request.args.get('page',1,type=int)
-    microblogs = Microblog.query.order_by(Microblog.timestamp.desc()).paginate(
-        page,current_app.config['POSTS_PER_PAGE'],False)
-    next_url = url_for('main.explore',page=microblogs.next_num) \
-        if microblogs.has_next else None
-    prev_url = url_for('main.explore',page=microblogs.prev_num) \
-        if microblogs.has_prev else None
-    return render_template('microblog/explore.html',title='Explore',microblogs=microblogs.items,next_url=next_url,prev_url=prev_url)
+    pagination = Microblog.query.order_by(Microblog.timestamp.desc()).paginate(
+        page,current_app.config['POSTS_PER_PAGE'])
+    microblogs = pagination.items
+    return render_template('microblog/explore.html', microblogs=microblogs, pagination=pagination)
     
 
 @bp.route('/blog/new',methods=['GET','POST'])
@@ -84,7 +77,7 @@ def new_blog():
             db.session.commit()
             flash('Blog created!')
             return redirect(url_for('main.manage_blog'))
-        return render_template('admin/new_blog.html',form=form)
+        return render_template('blog/new_blog.html',form=form)
     else:
         flash('Your should create your categories first.')
         return redirect(url_for('main.new_category'))
@@ -95,33 +88,35 @@ def new_blog():
 def show_blog(blog_id):
     blog = Blog.query.filter_by(id=blog_id,is_released=True).first_or_404()
     is_favourited = current_user.is_favourited_blog(blog)
-    return render_template('blog/blog.html', blog=blog,is_favourited=is_favourited)
+    is_author = blog.is_author(current_user)
+    categories = blog.get_author_categories()
+    links = blog.get_author_links()
+    return render_template('blog/blog.html', blog=blog, is_favourited=is_favourited, is_author=is_author, categories=categories, links=links)
     
 
 @bp.route('/blog/manage')
 @login_required
 def manage_blog():
     page = request.args.get('page', 1, type=int)
-    blogs = Blog.query.filter_by(author=current_user).order_by(Blog.timestamp.desc()).paginate(page, current_app.config['POSTS_PER_PAGE'],False)
-    next_url = url_for('main.index',page=blogs.next_num) \
-        if blogs.has_next else None
-    prev_url = url_for('main.index',page=blogs.prev_num) \
-        if blogs.has_prev else None
-    return render_template('admin/manage_blog.html', page=page, blogs=blogs.items,next_url=next_url,prev_url=prev_url)
+    #pagination = Blog.query.filter_by(author=current_user).order_by(Blog.timestamp.desc()).paginate(page, current_app.config['POSTS_PER_PAGE'])
+    pagination = current_user.get_allblogs().paginate(page, current_app.config['POSTS_PER_PAGE'])
+    blogs = pagination.items
+    return render_template('blog/manage_blog.html', blogs=blogs, pagination=pagination, page=page)
     
 
-@bp.route('/blog/<int:blog_id>/release', methods=['POST'])
+@bp.route('/blog/<int:blog_id>/release', methods=['GET', 'POST'])
 @login_required
 def release_blog(blog_id):
     blog = Blog.query.filter_by(id=blog_id,author=current_user).first_or_404()
 
     if blog.is_released:
         blog.is_released = False
+        db.session.commit()
         flash('Blog drafted.', 'success')
     else:
         blog.is_released = True
+        db.session.commit()
         flash('Blog released.', 'success')
-    db.session.commit()
     return redirect_back()
 
 
@@ -141,10 +136,10 @@ def edit_blog(blog_id):
     form.title.data = blog.title
     form.body.data = blog.body
     form.category.data = blog.category_id
-    return render_template('admin/edit_blog.html', form=form)
+    return render_template('blog/edit_blog.html', form=form)
     
 
-@bp.route('/blog/<int:blog_id>/delete', methods=['POST'])
+@bp.route('/blog/<int:blog_id>/delete', methods=['GET','POST'])
 @login_required
 def delete_blog(blog_id):
     blog = Blog.query.filter_by(id=blog_id,author=current_user).first_or_404()
@@ -163,23 +158,30 @@ def uploaded_files(filename):
 @bp.route('/upload', methods=['POST'])
 def upload():
     f = request.files.get('upload')
-    extension = f.filename.split('.')[1].lower()
-    if extension not in ['jpg', 'gif', 'png', 'jpeg']:
-        return upload_fail(message='Image only!')
-    f.save(os.path.join(current_app.config['UPLOADED_PATH'], time.strftime('%Y%m%d%H%M%S',time.localtime()) + '_' + f.filename))
-    url = url_for('main.uploaded_files', filename=time.strftime('%Y%m%d%H%M%S',time.localtime()) + '_' + f.filename)
-    return upload_success(url=url)
+    #extension = f.filename.split('.')[1].lower()
+    #if extension not in ['jpg', 'gif', 'png', 'jpeg']:
+    #    return upload_fail(message='Image only!')
+    if not allowed_file(f.filename):
+        return upload_fail('Image only!')
+
+    filename_modified = time.strftime('%Y%m%d%H%M%S',time.localtime()) + '_' + f.filename
+    #f.save(os.path.join(current_app.config['UPLOADED_PATH'], time.strftime('%Y%m%d%H%M%S',time.localtime()) + '_' + f.filename))
+    #url = url_for('main.uploaded_files', filename=time.strftime('%Y%m%d%H%M%S',time.localtime()) + '_' + f.filename)
+    f.save(os.path.join(current_app.config['UPLOADED_PATH'],filename_modified))
+    url = url_for('main.uploaded_files', filename=filename_modified)
+    return upload_success(url,filename_modified)
 
 
 @bp.route('/category/<int:category_id>')
+@login_required
 def show_category(category_id):
     category = Category.query.get_or_404(category_id)
     page = request.args.get('page', 1, type=int)
-    per_page = current_app.config['POSTS_PER_PAGE']
-    pagination = Blog.query.with_parent(category).order_by(Blog.timestamp.desc()).paginate(page, per_page)
+    pagination = category.get_releasedblogs().paginate(page, current_app.config['POSTS_PER_PAGE'])
     blogs = pagination.items
-    return render_template('blog/category.html', category=category, pagination=pagination, blogs=blogs)
-
+    categories = category.get_author_categories()
+    links = category.get_author_links()
+    return render_template('blog/category.html', category=category, blogs=blogs, pagination=pagination, categories=categories, links=links)
 
 @bp.route('/category/new', methods=['GET', 'POST'])
 @login_required
@@ -188,27 +190,24 @@ def new_category():
 
     if form.validate_on_submit():
         name = form.name.data
-        author = current_user
-        category = Category(name=name,author=author)
+        author = current_user.id
+        category = Category(name=name,user_id=author)
         db.session.add(category)
         if current_user.has_categories is not True:
             current_user.has_categories = True
         db.session.commit()
         flash('Category created.', 'success')
         return redirect(url_for('main.manage_category'))
-    return render_template('admin/new_category.html', form=form)
+    return render_template('blog/new_category.html', form=form)
     
 
 @bp.route('/category/manage')
 @login_required
 def manage_category():
     page = request.args.get('page',1,type=int)
-    categories = Category.query.filter_by(author=current_user).order_by(Category.id).paginate(page, current_app.config['POSTS_PER_PAGE'],False)
-    next_url = url_for('main.index',page=categories.next_num) \
-        if categories.has_next else None
-    prev_url = url_for('main.index',page=categories.prev_num) \
-        if categories.has_prev else None
-    return render_template('admin/manage_category.html',categories=categories.items, page=page,next_url=next_url,prev_url=prev_url)
+    pagination = Category.query.filter_by(author=current_user).order_by(Category.id).paginate(page, current_app.config['POSTS_PER_PAGE'])
+    categories = pagination.items
+    return render_template('blog/manage_category.html',categories=categories, pagination=pagination)
     
 
 @bp.route('/category/<int:category_id>/edit', methods=['GET', 'POST'])
@@ -223,20 +222,23 @@ def edit_category(category_id):
         flash('Category updated.', 'success')
         return redirect(url_for('main.manage_category'))
     form.name.data = category.name
-    return render_template('admin/edit_category.html', form=form)
+    return render_template('blog/edit_category.html', form=form)
 
 
-@bp.route('/category/<int:category_id>/delete', methods=['POST'])
+@bp.route('/category/<int:category_id>/delete', methods=['GET','POST'])
 @login_required
 def delete_category(category_id):
     category = Category.query.filter_by(id=category_id,author=current_user).first_or_404()
-    category.delete()
-    if Category.query.filter_by(author=current_user).first() is None:
-        current_user.has_categories = False
-        db.session.commit()
-        flash('Category deleted,and you have no category now.')
-    else: 
-        flash('Category deleted.', 'success')
-    return redirect(url_for('main.manage_category'))
-    
+
+    if category.is_blogsbelonged():
+        flash('Can not delete the category with blogs undered')
+    else:
+        category.delete()
+        if Category.query.filter_by(author=current_user).first() is None:
+            current_user.has_categories = False
+            db.session.commit()
+            flash('Category deleted,and you have no category now.')
+        else:
+            flash('Category deleted.', 'success')
+    return redirect_back()
